+++ Cargo.toml
@@ -12,11 +12,10 @@ epoll = ">=4.0.1"
 lazy_static = "1.4.0"
 libc = "0.2.66"
 log = { version = "0.4.10", features = ["std"] }
-net_gen = { path = "net_gen" }
-net_util = { path = "net_util" }
 qcow = { path = "qcow" }
 vhost_user_backend = { path = "vhost_user_backend"}
 vhost_user_fs = { path = "vhost_user_fs"}
+vhost_user_net = { path = "vhost_user_net"}
 virtio-bindings = "0.1.0"
 vmm = { path = "vmm" }
 vm-device = { path = "vm-device" }
@@ -58,6 +57,7 @@ members = [
     "vm-device",
     "vhost_user_backend",
     "vhost_user_fs",
+    "vhost_user_net",
     "vfio",
     "net_util",
     "acpi_tables",
