+++ vhost_user_net/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "vhost_user_net"
+version = "0.1.0"
+authors = ["The Cloud Hypervisor Authors"]
+edition = "2018"
+
+[dependencies]
+bitflags = "1.1.0"
+epoll = ">=4.0.1"
+libc = "0.2.65"
+log = "0.4.8"
+net_util = { path = "../net_util" }
+vhost_user_backend = { path = "../vhost_user_backend" }
+vhost_rs = { path = "../vhost_rs" }
+virtio-bindings = "0.1.0"
+vm-memory = { git = "https://github.com/rust-vmm/vm-memory" }
+vm-virtio = { path = "../vm-virtio" }
+vmm-sys-util = ">=0.3.1"
